[Mips] Explicitly cast ulittle32_t to the uint32_t to fix Visual Studio compile error.
llvm-svn: 197344

+++ lld/lib/ReaderWriter/ELF/Mips/MipsRelocationHandler.cpp
@@ -21,7 +21,7 @@ namespace {
 
 inline void applyReloc(uint8_t *location, uint32_t result) {
   auto target = reinterpret_cast<llvm::support::ulittle32_t *>(location);
-  *target = result | *target;
+  *target = result | uint32_t(*target);
 }
 
 /// \brief Calculate AHL value combines addends from 'hi' and 'lo' relocations.
